+++ trunk/webEdition/liveUpdate/sqldumps/tblnavigationrules.sql
@@ -4,9 +4,9 @@ CREATE TABLE ###TBLPREFIX###tblnavigationrules (
   NavigationID int unsigned NOT NULL default '0',
   SelectionType varchar(16) NOT NULL default '',
   FolderID int unsigned NOT NULL default '0',
-  DoctypeID int unsigned NOT NULL default '0',
+  DoctypeID mediumint unsigned NOT NULL default '0',
   Categories text NOT NULL,
-  ClassID int unsigned NOT NULL default '0',
+  ClassID smallint unsigned NOT NULL default '0',
   WorkspaceID int unsigned NOT NULL default '0',
   PRIMARY KEY  (ID)
 ) ENGINE=MyISAM;
